disassembler

Struct Disassembler

pub struct Disassembler {
    pub(crate) parser: Parser,
    pub(crate) cursor: usize,
    pub(crate) functions: Vec<Function>,
}

Fields
parser: Parsercursor: usizefunctions: Vec<Function>
Implementations

impl Disassembler

pub fn new(path: impl AsRef<Path>, nls: Nls) -> Result<Self>

pub fn get_parser(&self) -> &Parser

pub fn get_pc(&self) -> usize

pub fn nop(&mut self) -> Result<()>

0x00 nop instruction nop, no operation

pub fn init_stack(&mut self, parser: &mut Parser) -> Result<()>

0x01 init stack instruction initialize the local routine stack, as well as the post-phase of perforimg call instruction or launching a new routine

pub fn call(&mut self, parser: &mut Parser) -> Result<()>

0x02 call instruction call a routine

pub fn syscall(&mut self, parser: &mut Parser) -> Result<()>

0x03 syscall call a system call

pub fn ret(&mut self) -> Result<()>

0x04 ret instruction return from a routine

pub fn retv(&mut self) -> Result<()>

0x05 retv instruction return from a routine with a value

pub fn jmp(&mut self, parser: &mut Parser) -> Result<()>

0x06 jmp instruction jump to the address

pub fn jz(&mut self, parser: &mut Parser) -> Result<()>

0x07 jz instruction jump to the address if the top of the stack is zero

pub fn push_nil(&mut self) -> Result<()>

0x08 push nil push a nil value onto the stack

pub fn push_true(&mut self) -> Result<()>

0x09 push true push a true value onto the stack

pub fn push_i32(&mut self, parser: &mut Parser) -> Result<()>

0x0A push i32 push an i32 value onto the stack

pub fn push_i16(&mut self, parser: &mut Parser) -> Result<()>

0x0B push i16 push an i16 value onto the stack

pub fn push_i8(&mut self, parser: &mut Parser) -> Result<()>

0x0C push i8 push an i8 value onto the stack

pub fn push_f32(&mut self, parser: &mut Parser) -> Result<()>

0x0D push f32 push an f32 value onto the stack

pub fn push_string(&mut self, parser: &mut Parser) -> Result<()>

0x0E push string push a string onto the stack

pub fn push_global(&mut self, parser: &mut Parser) -> Result<()>

0x0F push global push a global variable onto the stack

pub fn push_stack(&mut self, parser: &mut Parser) -> Result<()>

0x10 push stack push a stack variable onto the stack

pub fn push_global_table(&mut self, parser: &mut Parser) -> Result<()>

0x11 push global table push a value than stored in the global table by immediate key onto the stack we assume that if any failure occurs, such as the key not found, we will push a nil value onto the stack for compatibility reasons.

pub fn push_local_table(&mut self, parser: &mut Parser) -> Result<()>

0x12 push local table push a value than stored in the local table by key onto the stack

pub fn push_top(&mut self) -> Result<()>

0x13 push top push the top of the stack onto the stack

pub fn push_return_value(&mut self) -> Result<()>

0x14 push return value push the return value onto the stack

pub fn pop_global(&mut self, parser: &mut Parser) -> Result<()>

0x15 pop global pop the top of the stack and store it in the global table

pub fn local_copy(&mut self, parser: &mut Parser) -> Result<()>

0x16 local copy copy the top of the stack to the local variable

pub fn pop_global_table(&mut self, parser: &mut Parser) -> Result<()>

0x17 pop global table pop the top of the stack and store it in the global table by key

pub fn pop_local_table(&mut self, parser: &mut Parser) -> Result<()>

0x18 pop local table pop the top of the stack and store it in the local table by key

pub fn neg(&mut self) -> Result<()>

0x19 neg negate the top of the stack, only works for integers and floats

pub fn add(&mut self) -> Result<()>

0x1A add add the top two values on the stack

pub fn sub(&mut self) -> Result<()>

0x1B sub subtract the top two values on the stack

pub fn mul(&mut self) -> Result<()>

0x1C mul multiply the top two values on the stack

pub fn div(&mut self) -> Result<()>

0x1D div divide the top two values on the stack

pub fn modulo(&mut self) -> Result<()>

0x1E modulo modulo the top two values on the stack

pub fn bittest(&mut self) -> Result<()>

0x1F bittest test with the top two values on the stack

pub fn and(&mut self) -> Result<()>

0x20 and push true if both the top two values on the stack are none-nil

pub fn or(&mut self) -> Result<()>

0x21 or push true if either of the top two values on the stack is none-nil

pub fn sete(&mut self) -> Result<()>

0x22 sete set the top of the stack to true if the top two values on the stack are equal

pub fn setne(&mut self) -> Result<()>

0x23 setne set the top of the stack to true if the top two values on the stack are not equal

pub fn setg(&mut self) -> Result<()>

0x24 setg set the top of the stack to true if the top two values on the stack are greater

pub fn setle(&mut self) -> Result<()>

0x25 setle set the top of the stack to true if the top two values on the stack are less or equal

pub fn setl(&mut self) -> Result<()>

0x26 setl set the top of the stack to true if the top two values on the stack are less

pub fn setge(&mut self) -> Result<()>

0x27 setge set the top of the stack to true if the top two values on the stack are greater or equal

pub(crate) fn disassemble_opcode(&mut self, parser: &mut Parser) -> Result<()>

pub fn disassemble(&mut self) -> Result<()>

pub fn write_insts(&self, path: impl AsRef<Path>) -> Result<()>